segundo commit

+++ app/Http/Controllers/LivroController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Livro;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class LivroController extends Controller
 {
@@ -15,6 +16,8 @@ class LivroController extends Controller
     
     public function create()
     {
+        //unico defeito é esse pc da ulbra nao ter o node pra testar a autenticação
+        Gate::authorize('create', Livro::class);
         return view('livros.create');
     }
     
@@ -26,6 +29,7 @@ class LivroController extends Controller
     
     public function edit($id)
     {
+        Gate::authorize('update', Livro::class);
         $livro = Livro::findOrFail($id);
         return view('livros.edit', compact('livro'));
     }
@@ -39,6 +43,7 @@ class LivroController extends Controller
     
     public function destroy($id)
     {
+        Gate::authorize('delete', Livro::class);
         $livro = Livro::findOrFail($id);
         $livro->delete();
         return redirect('livros')->with('success', 'Livro deleted successfully.');
